[hy] adjust return type

+++ unitsconversion/src/main/java/name/hanyi/unitsconversion/UnitsConversion.java
@@ -7,9 +7,8 @@ public class UnitsConversion {
     public static final int PRECISION = 8;
     public static final double RATIO_POUND_TO_KILO = 0.45359237D;
 
-    public Double get(double number) {
+    public BigDecimal get(double number) {
         return BigDecimal.valueOf(number * RATIO_POUND_TO_KILO)
-                .setScale(PRECISION, RoundingMode.HALF_UP)
-                .doubleValue();
+                .setScale(PRECISION, RoundingMode.HALF_UP);
     }
 }
